requirements.txt
opencv-python==4.8.1.78
ultralytics==8.0.196
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
Pillow==10.0.0
